refactor: move lib to src
