+++ src/configue-core.js
