+++ test/builder.test.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const Configue = require('../lib/configue');
+const Configue = require('../src/configue');
 
 test('get is working fine as factory method', t => {
     const configue = Configue.get();
