+++ test/getters.getobject.test.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const Configue = require('../lib/configue');
+const Configue = require('../src/configue');
 
 test('basic getObject', t => {
     const configue = Configue({defaults: {A: '2', B: 42}});
